Add the InputStream and OutputStream classes to libdua

+++ src/parsing/ParserAssistant.cpp
@@ -28,9 +28,9 @@ if (num[0] == '0')                                            \
 return FUNC(num);
 
 int64_t ParserAssistant::get_i64(std::string num) { get_ixx(std::stoll); }
-int32_t ParserAssistant::get_i32(std::string num) { get_ixx(std::stoi); }
-int16_t ParserAssistant::get_i16(std::string num) { get_ixx(std::stoi); }
-int8_t  ParserAssistant::get_i8 (std::string num) { get_ixx(std::stoi); }
+int32_t ParserAssistant::get_i32(std::string num) { get_ixx(std::stoll); }
+int16_t ParserAssistant::get_i16(std::string num) { get_ixx(std::stoll); }
+int8_t  ParserAssistant::get_i8 (std::string num) { get_ixx(std::stoll); }
 
 TranslationUnitNode* ParserAssistant::construct_result()
 {
